+++ src/app/post-ad-component/post-ad-component.component.css
@@ -7,14 +7,14 @@
     font-weight: lighter;
     box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
 }
- 
+
 .button-container {
     margin: 0 auto;
     width: 100%;
     display: flex;
     justify-content: space-evenly;
 }
- 
+
 #post-ad-container .button-container .btn {
     width: 40%;
     border-radius: 15px;
@@ -45,7 +45,6 @@
     border-radius: 100px;
     background: white;
     padding: 0;
-
 }
 
 .closeButton .fa {
@@ -55,7 +54,6 @@
     padding-right: 3px;
     padding-top: 2px;
     padding-bottom: 2px;
-
 }
 
 #uploadImage img {
@@ -71,11 +69,12 @@
     top: 0;
     right: 0;
     cursor: pointer;
-
 }
-.btn{
+
+.btn {
     padding: 0;
 }
+
 .addEditButton {
     z-index: 3;
     background: white;
@@ -86,8 +85,8 @@
     transform: translate(-50%, -50%);
     padding: 0px;
 }
+
 .progressBar {
-    
     background-color: blue;
     border-radius: 15px;
     border-radius: 15px;
@@ -97,29 +96,33 @@
     left: 50%;
     transform: translate(-50%, -50%);
 }
+
 .spinner-border {
     position: absolute;
     top: 40%;
     left: 43%;
 }
 
-.img-error{
+.img-error {
     font-size: 12px;
 }
 
+.ad-action-btn:hover {
+    cursor: pointer;
+}
 
 /*DataPicker code - see below- >*/
- 
+
 input[type="date"] {
     position: relative;
 }
- 
+
 input[type="date"]:after {
     content: "\25BC";
     color: #555;
     padding: 0 5px;
 }
- 
+
 input[type="date"]::-webkit-calendar-picker-indicator {
     position: absolute;
     top: 0;
@@ -131,7 +134,20 @@ input[type="date"]::-webkit-calendar-picker-indicator {
     color: transparent;
     background: transparent;
 }
- 
+
+input[type="date"]::-webkit-clear-button, input[type="date"]::-webkit-inner-spin-button {
+    display: none;
+    -webkit-appearance: none;
+    margin: 0;
+}
+
+/* Number */
+
+input[type="number"]::-webkit-outer-spin-button, input[type="number"]::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+    margin: 0;
+}
+
 @media only screen and (max-width: 700px) {
     #post-ad-container {
         margin: 10px auto;
