+++ src/app/services/product-mock.service.ts
@@ -4,10 +4,15 @@ import { IProductService } from '../models/i-product-service';
 import { Observable, of } from 'rxjs';
 import { GetProductDetailsResponse } from '../models/get-product-details-response';
 import { GetProductsListResponse } from '../models/get-products-list-response';
+import { environment } from 'src/environments/environment';
 
 export class ProductMockService implements IProductService {
 
-  constructor() { }
+  productsList: Product[];
+
+  constructor() {
+    this.setDummyProductList();
+  }
 
   AddProduct(product: Product): Observable<Product> {
     product.id = "P123";
@@ -20,6 +25,21 @@ export class ProductMockService implements IProductService {
     // set other details coming from web
     return of(product);
   }
+
+  updateProduct(product: Product): Observable<Product> {
+    let index;
+
+    for (index = 0; index < this.productsList.length; index++) {
+      if (this.productsList[index].id == product.id) {
+        this.productsList[index] = product;
+        break;
+      }
+    }
+    console.log(this.productsList[index]);
+    // set other details coming from web
+    return of(this.productsList[index]);
+  }
+
   GetMockPreview(product: Product): GetProductDetailsResponse {
     let sellerObj = new Seller();
     sellerObj.id = "1";
@@ -30,7 +50,7 @@ export class ProductMockService implements IProductService {
     return productpreviewObj;
   }
 
-  getDummyProductList(): Product[] {
+  setDummyProductList() {
     let product1 = new Product();
     product1.id = "1";
     product1.pickupAddress.line1 = "Satara";
@@ -228,11 +248,14 @@ export class ProductMockService implements IProductService {
     product13.price.isNegotiable = true;
     product13.images.push('https://images.pexels.com/photos/170811/pexels-photo-170811.jpeg?auto=compress&cs=tinysrgb&h=750&w=1260');
 
-    let itemList: Product[] = [
+    this.productsList = [
       product1, product2, product3, product4, product5, product6, product7,
       product8, product9, product10, product11, product12, product13
     ];
-    return itemList;
+  }
+
+  getDummyProductList(): Product[] {
+    return this.productsList;
   }
 
   getDummyProductDetails(productId): GetProductDetailsResponse {
@@ -281,4 +304,97 @@ export class ProductMockService implements IProductService {
     getProductsListResponse.products = this.getDummyProductList();
     return of(getProductsListResponse);
   }
+
+  getDummyActiveUserProducts(): Product[] {
+    let product1 = new Product();
+    product1.id = "3";
+    product1.pickupAddress.line1 = "Nashik Road";
+    product1.pickupAddress.city = "Nashik ";
+    product1.pickupAddress.state = "Maharshtra";
+    product1.name = "Fun stories about car and kids";
+    product1.category = "cars";
+    product1.description = "Just A Photo";
+    product1.postDateTime = new Date();
+    product1.heroImage = "http://i.imgur.com/REM4kQUg.jpg";
+    product1.price.money.amount = 730000;
+    product1.price.money.currency = "INR";
+    product1.price.isNegotiable = true;
+    product1.images.push('http://i.imgur.com/REM4kQUg.jpg');
+
+    let product2 = new Product();
+    product2.id = "4";
+    product2.pickupAddress.line1 = "Kothrud Paschima Nagri";
+    product2.pickupAddress.city = "Pune";
+    product2.pickupAddress.state = "Maharshtra";
+    product2.name = "Hyundai I20 i20 Asta 1.2, 2014, Petrol";
+    product2.category = "cars";
+    product2.description = "Maruti Suzuki India Limited, formerly known as Maruti Udyog Limited, is an automobile manufacturer in India. It is a 56.21% owned subsidiary of the Japanese car and motorcycle manufacturer Suzuki Motor Corporation. As of July 2018, it had a market share of 53% of the Indian passenger car market.";
+    product2.postDateTime = new Date();
+    product2.heroImage = "https://i.ytimg.com/vi/dsWxMoh3_50/maxresdefault.jpg";
+    product2.price.money.amount = 6800000;
+    product2.price.money.currency = "INR";
+    product2.price.isNegotiable = true;
+    product2.images.push('https://i.ytimg.com/vi/dsWxMoh3_50/maxresdefault.jpg');
+
+    let itemList: Product[] = [product1, product2];
+    return itemList;
+  }
+
+  getDummyInactiveUserProducts(): Product[] {
+    let product1 = new Product();
+    product1.id = "5";
+    product1.pickupAddress.line1 = "Baner";
+    product1.pickupAddress.city = "Pune";
+    product1.pickupAddress.state = "Maharshtra";
+    product1.name = "Maruti Suzuki Wagon R LXI, 2016, CNG & Hybrids";
+    product1.category = "cars";
+    product1.description = "Most Beautiful Car in The World";
+    product1.postDateTime = new Date();
+    product1.heroImage = "https://image.shutterstock.com/image-photo/big-small-ripe-red-strawberry-260nw-642144988.jpg";
+    product1.price.money.amount = 460000;
+    product1.price.money.currency = "INR";
+    product1.price.isNegotiable = true;
+    product1.images.push('https://image.shutterstock.com/image-photo/big-small-ripe-red-strawberry-260nw-642144988.jpg');
+
+    let product2 = new Product();
+    product2.id = "6";
+    product2.pickupAddress.line1 = "Dombivali";
+    product2.pickupAddress.city = "Mumbai";
+    product2.pickupAddress.state = "Maharashtra";
+    product2.name = "Volkswagen Vento Highline Diesel AT, 2016, Diesel";
+    product2.category = "cars";
+    product2.description = "New Car";
+    product2.postDateTime = new Date();
+    product2.heroImage = "https://cdn.vox-cdn.com/thumbor/qW3lRejqR7xJAG_FTdRyxjfbalA=/0x0:2040x1360/1200x800/filters:focal(857x517:1183x843)/cdn.vox-cdn.com/uploads/chorus_image/image/65408496/tesla_model_3_0638.0.jpg";
+    product2.price.money.amount = 8600000;
+    product2.price.money.currency = "INR";
+    product2.price.isNegotiable = true;
+    product2.images.push('https://cdn.vox-cdn.com/thumbor/qW3lRejqR7xJAG_FTdRyxjfbalA=/0x0:2040x1360/1200x800/filters:focal(857x517:1183x843)/cdn.vox-cdn.com/uploads/chorus_image/image/65408496/tesla_model_3_0638.0.jpg');
+
+    let itemList: Product[] = [product1, product2];
+    return itemList;
+  }
+
+  getActiveUserProducts(
+    userId: string
+  ): Observable<GetProductsListResponse> {
+    if (userId != environment.userSetting.userId) {
+      return null;
+    }
+    var getProductsListResponse = new GetProductsListResponse();
+    getProductsListResponse.products = this.getDummyActiveUserProducts();
+    return of(getProductsListResponse);
+  }
+
+  getInactiveUserProducts(
+    userId: string
+  ): Observable<GetProductsListResponse> {
+    if (userId != environment.userSetting.userId) {
+      return null;
+    }
+    var getProductsListResponse = new GetProductsListResponse();
+    getProductsListResponse.products = this.getDummyInactiveUserProducts();
+    return of(getProductsListResponse);
+  }
+
 }
