+++ src/app/services/user/user.service.ts
@@ -3,14 +3,14 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { environment } from 'src/environments/environment';
 import { UserMockService } from './user-mock.service';
 import { GetUserProfileResponse } from 'src/app/models/user/get-user-profile-response';
-import { Observable, of, BehaviorSubject, } from 'rxjs';
+import { Observable, BehaviorSubject } from 'rxjs';
 import { IUserService } from 'src/app/models/i-user-service';
-import { GetProductsListResponse } from 'src/app/models/get-products-list-response';
 import { Product } from 'src/app/models/product';
 
 @Injectable({
   providedIn: 'root'
 })
+
 export class UserService implements IUserService {
   userId: string;
   userMockService: UserMockService;
@@ -40,33 +40,7 @@ export class UserService implements IUserService {
     }
   }
 
-  getActiveUserProducts(
-    userId: string
-  ): Observable<GetProductsListResponse> {
-    if (environment.isMockingEnabled) {
-      return this.userMockService.getActiveUserProducts(userId);
-    } else {
-      let getProductListUrl: string = this.getUrl(environment.userSetting.profile) + userId + environment.userSetting.activeAds;
-      return this.http.get<GetProductsListResponse>(getProductListUrl, {
-        headers: this.headers
-      });
-    }
-  }
-
-  getInactiveUserProducts(
-    userId: string
-  ): Observable<GetProductsListResponse> {
-    if (environment.isMockingEnabled) {
-      return this.userMockService.getInactiveUserProducts(userId);
-    } else {
-      let getProductListUrl: string = this.getUrl(environment.userSetting.inactiveAds) + userId + environment.userSetting.inactiveAds;
-      return this.http.get<GetProductsListResponse>(getProductListUrl, {
-        headers: this.headers
-      });
-    }
-  }
-
-  private getUrl(path: string): string {
+  getUrl(path: string): string {
     return
     environment.userSetting.baseUrl +
       environment.version +
